[REFACTOR] Reorganize renderer and style computation code
Rename MxGraphRenderer into BpmnRenderer: better naming
Extract style computation out of StyleConfigurator, in a new StyleComputer class
  - this prevents the renderer to depend on a configuration class (no need for
  the mxGraph instance)
  - the computation code doesn't depend on the configuration code, it only uses
  it. So show it by splitting the code into 2 classes
