+++ src/component/mxgraph/config/StyleConfigurator.ts
@@ -275,135 +275,4 @@ export default class StyleConfigurator {
     this.configureSequenceFlowStyles();
     this.configureAssociationFlowStyles();
   }
-
-  computeStyle(bpmnCell: Shape | Edge, labelBounds: Bounds): string {
-    const styles: string[] = [bpmnCell.bpmnElement.kind as string];
-
-    let shapeStyleValues;
-    if (bpmnCell instanceof Shape) {
-      shapeStyleValues = this.computeShapeStyle(bpmnCell);
-    } else {
-      styles.push(...this.computeEdgeStyle(bpmnCell));
-      shapeStyleValues = new Map<string, string | number>();
-    }
-
-    const fontStyleValues = this.computeFontStyleValues(bpmnCell);
-    const labelStyleValues = this.computeLabelStyleValues(bpmnCell, labelBounds);
-
-    return [] //
-      .concat([...styles])
-      .concat([...shapeStyleValues, ...fontStyleValues, ...labelStyleValues].filter(([, v]) => v && v != 'undefined').map(([key, value]) => key + '=' + value))
-      .join(';');
-  }
-
-  private computeShapeStyle(shape: Shape): Map<string, string | number> {
-    const styleValues = new Map<string, string | number>();
-    const bpmnElement = shape.bpmnElement;
-
-    if (bpmnElement instanceof ShapeBpmnEvent) {
-      styleValues.set(StyleIdentifier.BPMN_STYLE_EVENT_KIND, bpmnElement.eventKind);
-
-      if (bpmnElement instanceof ShapeBpmnBoundaryEvent || (bpmnElement instanceof ShapeBpmnStartEvent && bpmnElement.isInterrupting !== undefined)) {
-        styleValues.set(StyleIdentifier.BPMN_STYLE_IS_INTERRUPTING, String(bpmnElement.isInterrupting));
-      }
-    } else if (bpmnElement instanceof ShapeBpmnActivity) {
-      if (bpmnElement instanceof ShapeBpmnSubProcess) {
-        styleValues.set(StyleIdentifier.BPMN_STYLE_SUB_PROCESS_KIND, bpmnElement.subProcessKind);
-      } else if (bpmnElement.kind === ShapeBpmnElementKind.TASK_RECEIVE) {
-        styleValues.set(StyleIdentifier.BPMN_STYLE_INSTANTIATING, String(bpmnElement.instantiate));
-      }
-
-      const markers: ShapeBpmnMarkerKind[] = bpmnElement.markers;
-      if (markers.length > 0) {
-        styleValues.set(StyleIdentifier.BPMN_STYLE_MARKERS, markers.join(','));
-      }
-    } else if (ShapeUtil.isPoolOrLane(bpmnElement.kind)) {
-      // mxgraph.mxConstants.STYLE_HORIZONTAL is for the label
-      // In BPMN, isHorizontal is for the Shape
-      styleValues.set(mxgraph.mxConstants.STYLE_HORIZONTAL, shape.isHorizontal ? '0' : '1');
-    } else if (bpmnElement instanceof ShapeBpmnEventBasedGateway) {
-      styleValues.set(StyleIdentifier.BPMN_STYLE_INSTANTIATING, String(bpmnElement.instantiate));
-      styleValues.set(StyleIdentifier.BPMN_STYLE_EVENT_BASED_GATEWAY_KIND, String(bpmnElement.gatewayKind));
-    }
-
-    return styleValues;
-  }
-
-  private computeEdgeStyle(edge: Edge): string[] {
-    const styles: string[] = [];
-
-    const bpmnElement = edge.bpmnElement;
-    if (bpmnElement instanceof SequenceFlow) {
-      styles.push(bpmnElement.sequenceFlowKind);
-    }
-    if (bpmnElement instanceof AssociationFlow) {
-      styles.push(bpmnElement.associationDirectionKind);
-    }
-
-    return styles;
-  }
-
-  private computeFontStyleValues(bpmnCell: Shape | Edge): Map<string, string | number> {
-    const styleValues = new Map<string, string | number>();
-
-    const font = bpmnCell.label?.font;
-    if (font) {
-      styleValues.set(mxgraph.mxConstants.STYLE_FONTFAMILY, font.name);
-      styleValues.set(mxgraph.mxConstants.STYLE_FONTSIZE, font.size);
-      styleValues.set(mxgraph.mxConstants.STYLE_FONTSTYLE, StyleConfigurator.getFontStyleValue(font));
-    }
-
-    return styleValues;
-  }
-
-  private computeLabelStyleValues(bpmnCell: Shape | Edge, labelBounds: Bounds): Map<string, string | number> {
-    const styleValues = new Map<string, string | number>();
-
-    const bpmnElement = bpmnCell.bpmnElement;
-    if (labelBounds) {
-      styleValues.set(mxgraph.mxConstants.STYLE_VERTICAL_ALIGN, mxgraph.mxConstants.ALIGN_TOP);
-      if (bpmnCell.bpmnElement.kind != ShapeBpmnElementKind.TEXT_ANNOTATION) {
-        styleValues.set(mxgraph.mxConstants.STYLE_ALIGN, mxgraph.mxConstants.ALIGN_CENTER);
-      }
-
-      if (bpmnCell instanceof Shape) {
-        // arbitrarily increase width to relax too small bounds (for instance for reference diagrams from miwg-test-suite)
-        styleValues.set(mxgraph.mxConstants.STYLE_LABEL_WIDTH, labelBounds.width + 1);
-        // align settings
-        styleValues.set(mxgraph.mxConstants.STYLE_LABEL_POSITION, mxgraph.mxConstants.ALIGN_TOP);
-        styleValues.set(mxgraph.mxConstants.STYLE_VERTICAL_LABEL_POSITION, mxgraph.mxConstants.ALIGN_LEFT);
-      }
-    }
-    // when no label bounds, adjust the default style dynamically
-    else if (
-      bpmnCell instanceof Shape &&
-      (bpmnElement instanceof ShapeBpmnSubProcess || bpmnElement instanceof ShapeBpmnCallActivity) &&
-      !bpmnElement.markers.includes(ShapeBpmnMarkerKind.EXPAND)
-    ) {
-      styleValues.set(mxgraph.mxConstants.STYLE_VERTICAL_ALIGN, mxgraph.mxConstants.ALIGN_TOP);
-    }
-
-    return styleValues;
-  }
-
-  computeMessageFlowIconStyle(edge: Edge): string {
-    return `shape=${StyleIdentifier.BPMN_STYLE_MESSAGE_FLOW_ICON};${StyleIdentifier.BPMN_STYLE_IS_INITIATING}=${edge.messageVisibleKind}`;
-  }
-
-  private static getFontStyleValue(font: Font): number {
-    let value = 0;
-    if (font.isBold) {
-      value += mxgraph.mxConstants.FONT_BOLD;
-    }
-    if (font.isItalic) {
-      value += mxgraph.mxConstants.FONT_ITALIC;
-    }
-    if (font.isStrikeThrough) {
-      value += mxgraph.mxConstants.FONT_STRIKETHROUGH;
-    }
-    if (font.isUnderline) {
-      value += mxgraph.mxConstants.FONT_UNDERLINE;
-    }
-    return value;
-  }
 }
